Move the Sarvam realtime STT service into the provider's stt module
Each provider keeps its speech-to-text services in one module, the way
ElevenLabs and OpenAI hold their batch and realtime services together, and
its tests in one file alongside them.
