--- src/pipecat/services/sarvam/realtime_stt.py
@@ -1,904 +1,0 @@
-#
-# Copyright (c) 2024-2026, Daily
-#
-# SPDX-License-Identifier: BSD 2-Clause License
-#
-
-"""Sarvam realtime Speech-to-Text service implementation."""
-
-import asyncio
-import base64
-import json
-import time
-from collections.abc import AsyncGenerator, Awaitable, Callable
-from dataclasses import dataclass, field, fields
-from typing import Any, Literal
-from urllib.parse import urlencode
-
-from loguru import logger
-from websockets.asyncio.client import connect as websocket_connect
-from websockets.protocol import State
-
-from pipecat.frames.frames import (
-    CancelFrame,
-    EndFrame,
-    ErrorFrame,
-    Frame,
-    InterimTranscriptionFrame,
-    StartFrame,
-    STTMetadataFrame,
-    TranscriptionFrame,
-    UserStartedSpeakingFrame,
-    UserStoppedSpeakingFrame,
-    VADUserStartedSpeakingFrame,
-    VADUserStoppedSpeakingFrame,
-)
-from pipecat.processors.frame_processor import FrameDirection
-from pipecat.services.sarvam._sdk import sdk_headers
-from pipecat.services.settings import STTSettings
-from pipecat.services.stt_latency import SARVAM_TTFS_P99
-from pipecat.services.stt_service import WebsocketSTTService
-from pipecat.transcriptions.language import Language, resolve_language
-from pipecat.turns.user_turn_strategies import ExternalUserTurnStrategies
-from pipecat.utils.time import time_now_iso8601
-from pipecat.utils.tracing.service_decorators import traced_stt
-from pipecat.utils.types import NOT_GIVEN, NotGiven, assert_given, is_given
-
-_REALTIME_STT_URL = "wss://api.sarvam.ai/speech-to-text-realtime/ws"
-_REALTIME_MODEL = "saaras:v3-realtime"
-
-SUPPORTED_LANGUAGES = {
-    "auto",
-    "en-IN",
-    "hi-IN",
-    "bn-IN",
-    "kn-IN",
-    "ml-IN",
-    "mr-IN",
-    "or-IN",
-    "pa-IN",
-    "ta-IN",
-    "te-IN",
-    "gu-IN",
-    "as-IN",
-    "ur-IN",
-    "ne-IN",
-    "kok-IN",
-    "ks-IN",
-    "sd-IN",
-    "sa-IN",
-    "sat-IN",
-    "mni-IN",
-    "brx-IN",
-    "mai-IN",
-    "doi-IN",
-}
-SUPPORTED_STREAM_TYPES = {"fast", "balanced", "simulated"}
-SUPPORTED_ENDPOINTING = {"vad", "manual"}
-SUPPORTED_SAMPLE_RATES = {8000, 16000}
-SUPPORTED_MODES = {"transcribe", "translate", "verbatim", "translit", "codemix"}
-
-# Sarvam's `stream_type` selects the *server* flush profile; it says nothing
-# about how often the client should send. Audio goes out on a fixed cadence so
-# transcript latency doesn't scale with the chosen profile.
-_CLIENT_CHUNK_MS = 50
-
-# Fixed for the life of the socket — Sarvam only reads these from the connection
-# query string, so a change needs a new stream.
-_CONNECTION_ONLY_FIELDS = frozenset(
-    {
-        "sample_rate",
-        "return_timestamps",
-        "prefix_padding_ms",
-        "lid_gate_seconds",
-        "lid_confidence_threshold",
-    }
-)
-
-# Accepted in a `config.update` message. Of these, `language_code`,
-# `stream_type`, `mode`, `prompt`, and `endpointing` are boundary-gated: the
-# server defers them to the next utterance boundary.
-_RUNTIME_CONFIG_FIELDS = frozenset(
-    {
-        "language_code",
-        "stream_type",
-        "endpointing",
-        "mode",
-        "prompt",
-        "threshold",
-        "silence_duration_ms",
-        "min_speech_duration_ms",
-    }
-)
-
-# Only meaningful while the server is doing the endpointing.
-_VAD_TUNING_FIELDS = (
-    "threshold",
-    "prefix_padding_ms",
-    "silence_duration_ms",
-    "min_speech_duration_ms",
-)
-_SHORT_LANGUAGE_DEFAULTS = {
-    language.split("-", maxsplit=1)[0]: language
-    for language in SUPPORTED_LANGUAGES
-    if language != "auto"
-}
-
-
-def language_to_sarvam_realtime_language(language: Language) -> str:
-    """Convert a Language enum to Sarvam realtime's language code."""
-    language_map = {
-        Language.AS_IN: "as-IN",
-        Language.BN_IN: "bn-IN",
-        Language.EN_IN: "en-IN",
-        Language.GU_IN: "gu-IN",
-        Language.HI_IN: "hi-IN",
-        Language.KN_IN: "kn-IN",
-        Language.KOK_IN: "kok-IN",
-        Language.MAI_IN: "mai-IN",
-        Language.ML_IN: "ml-IN",
-        Language.MR_IN: "mr-IN",
-        Language.OR_IN: "or-IN",
-        Language.PA_IN: "pa-IN",
-        Language.SD_IN: "sd-IN",
-        Language.TA_IN: "ta-IN",
-        Language.TE_IN: "te-IN",
-    }
-    return resolve_language(language, language_map, use_base_code=False)
-
-
-@dataclass
-class SarvamRealtimeSTTSettings(STTSettings):
-    """Settings for SarvamRealtimeSTTService.
-
-    Parameters:
-        language_code: Sarvam realtime language code or ``auto``.
-        stream_type: Streaming cadence: ``fast``, ``balanced``, or ``simulated``.
-        endpointing: Turn endpointing mode: ``vad`` or ``manual``.
-        sample_rate: Declared input audio sample rate. ``None`` adopts the
-            pipeline's input rate.
-        mode: Realtime STT task mode.
-        prompt: Optional decoding prompt.
-        return_timestamps: Whether final transcripts should include segment offsets.
-        threshold: Optional VAD sensitivity threshold.
-        prefix_padding_ms: Optional VAD prefix padding.
-        silence_duration_ms: Optional silence duration for end-of-speech.
-        min_speech_duration_ms: Optional minimum speech duration.
-        lid_gate_seconds: Optional auto-LID gate duration.
-        lid_confidence_threshold: Optional auto-LID confidence threshold.
-    """
-
-    language_code: str | NotGiven = field(default_factory=lambda: NOT_GIVEN)
-    stream_type: str | NotGiven = field(default_factory=lambda: NOT_GIVEN)
-    endpointing: str | NotGiven = field(default_factory=lambda: NOT_GIVEN)
-    sample_rate: int | None | NotGiven = field(default_factory=lambda: NOT_GIVEN)
-    mode: Literal["transcribe", "translate", "verbatim", "translit", "codemix"] | NotGiven = field(
-        default_factory=lambda: NOT_GIVEN
-    )
-    prompt: str | None | NotGiven = field(default_factory=lambda: NOT_GIVEN)
-    return_timestamps: bool | NotGiven = field(default_factory=lambda: NOT_GIVEN)
-    threshold: float | None | NotGiven = field(default_factory=lambda: NOT_GIVEN)
-    prefix_padding_ms: int | None | NotGiven = field(default_factory=lambda: NOT_GIVEN)
-    silence_duration_ms: int | None | NotGiven = field(default_factory=lambda: NOT_GIVEN)
-    min_speech_duration_ms: int | None | NotGiven = field(default_factory=lambda: NOT_GIVEN)
-    lid_gate_seconds: float | None | NotGiven = field(default_factory=lambda: NOT_GIVEN)
-    lid_confidence_threshold: float | None | NotGiven = field(default_factory=lambda: NOT_GIVEN)
-
-
-class SarvamRealtimeSTTService(WebsocketSTTService):
-    """Sarvam realtime Speech-to-Text service.
-
-    Streams raw audio bytes to Sarvam's realtime websocket endpoint and maps
-    provider VAD and transcript events into Pipecat frames.
-
-    With the default ``endpointing="vad"`` the server drives turn boundaries. With
-    ``endpointing="manual"`` the pipeline drives them instead, so the pipeline
-    needs a turn strategy that emits ``VADUserStartedSpeakingFrame`` and
-    ``VADUserStoppedSpeakingFrame``; without one, Sarvam never receives a boundary
-    and never emits a final transcript.
-    """
-
-    Settings = SarvamRealtimeSTTSettings
-    _settings: Settings
-
-    def __init__(
-        self,
-        *,
-        api_key: str,
-        base_url: str = _REALTIME_STT_URL,
-        settings: Settings | None = None,
-        should_interrupt: bool = True,
-        ttfs_p99_latency: float | None = SARVAM_TTFS_P99,
-        **kwargs,
-    ):
-        """Initialize Sarvam realtime STT.
-
-        Args:
-            api_key: Sarvam API key.
-            base_url: Realtime STT websocket endpoint.
-            settings: Runtime-updatable realtime settings.
-            should_interrupt: Whether provider speech-start events should broadcast interruption.
-            ttfs_p99_latency: P99 latency from speech end to final transcript in seconds.
-            **kwargs: Additional arguments passed to :class:`WebsocketSTTService`.
-        """
-        settings_fields = {setting.name for setting in fields(self.Settings)}
-        direct_settings = sorted(settings_fields.intersection(kwargs))
-        if direct_settings:
-            names = ", ".join(direct_settings)
-            raise TypeError(
-                f"{names} must be passed via "
-                "settings=SarvamRealtimeSTTService.Settings(...), not as constructor kwargs"
-            )
-        if "reconnect_on_error" in kwargs:
-            raise TypeError(
-                "SarvamRealtimeSTTService does not support reconnect_on_error; "
-                "reconnection is always disabled"
-            )
-        default_settings = self.Settings(
-            model=_REALTIME_MODEL,
-            language=None,
-            language_code="en-IN",
-            stream_type="balanced",
-            endpointing="vad",
-            sample_rate=None,
-            mode="transcribe",
-            prompt=None,
-            return_timestamps=False,
-            threshold=None,
-            prefix_padding_ms=None,
-            silence_duration_ms=None,
-            min_speech_duration_ms=None,
-            lid_gate_seconds=None,
-            lid_confidence_threshold=None,
-        )
-        language_code_given = settings is not None and is_given(settings.language_code)
-        if settings is not None:
-            default_settings.apply_update(settings)
-
-        # An explicit language_code wins; otherwise derive it from `language`.
-        # `language` may still be a raw string here, since the base class only
-        # normalizes it once super().__init__() runs.
-        if not language_code_given:
-            language = _as_language(default_settings.language)
-            if language is not None:
-                default_settings.language_code = language_to_sarvam_realtime_language(language)
-
-        self._validate_settings(default_settings)
-
-        super().__init__(
-            sample_rate=assert_given(default_settings.sample_rate),
-            settings=default_settings,
-            ttfs_p99_latency=ttfs_p99_latency,
-            reconnect_on_error=False,
-            **kwargs,
-        )
-
-        self._api_key = api_key
-        self._base_url = base_url
-        self._should_interrupt = should_interrupt
-        self._receive_task: asyncio.Task | None = None
-        self._audio_buffer = bytearray()
-        self._request_id: str | None = None
-        self._provider_speech_active = False
-        self._speech_end_wall_time: float | None = None
-        self._speech_end_audio_position_s: float | None = None
-        self._audio_position_bytes = 0
-        # What the server is actually doing, which trails `settings.endpointing`
-        # until a `config.updated` acknowledges the switch.
-        self._effective_endpointing = assert_given(self._settings.endpointing)
-        self._pending_endpointing: str | None = None
-
-    def can_generate_metrics(self) -> bool:
-        """Check if this service can generate processing and usage metrics."""
-        return True
-
-    def service_metadata_frame(self) -> STTMetadataFrame:
-        """Request external turn strategies when Sarvam endpoints server-side.
-
-        With ``endpointing="vad"`` (the default) Sarvam's VAD decides turn
-        boundaries and emits ``UserStarted/StoppedSpeakingFrame``, so the user
-        aggregator defers to those rather than running local VAD/smart-turn. In
-        ``endpointing="manual"`` the pipeline supplies the boundaries, so the
-        defaults are left in place. Applied unless the user passed their own
-        ``user_turn_strategies``.
-        """
-        frame = super().service_metadata_frame()
-        if self._effective_endpointing == "vad":
-            frame.user_turn_strategies = ExternalUserTurnStrategies()
-        return frame
-
-    def language_to_service_language(self, language: Language) -> str:
-        """Convert a Language enum to Sarvam realtime's language code."""
-        return language_to_sarvam_realtime_language(language)
-
-    async def start(self, frame: StartFrame):
-        """Start the service and connect the websocket."""
-        await super().start(frame)
-        # The rate can come from the pipeline, so it is only known now. Report
-        # it rather than raise: `AIService._start` swallows exceptions, which
-        # would leave the service quietly dropping every audio chunk instead.
-        error = self._resolved_sample_rate_error()
-        if error:
-            await self.push_error(error)
-            return
-        await self._connect()
-
-    async def stop(self, frame: EndFrame):
-        """Stop the service and close the websocket."""
-        await super().stop(frame)
-        await self._disconnect()
-
-    async def cancel(self, frame: CancelFrame):
-        """Cancel the service and close the websocket."""
-        await super().cancel(frame)
-        await self._disconnect()
-
-    async def process_frame(self, frame: Frame, direction: FrameDirection):
-        """Process frames and send manual endpointing boundaries when configured."""
-        await super().process_frame(frame, direction)
-        if self._effective_endpointing != "manual":
-            return
-        if isinstance(frame, VADUserStartedSpeakingFrame):
-            await self._send_json({"event": "speech_start"})
-        elif isinstance(frame, VADUserStoppedSpeakingFrame):
-            # The tail of the turn must reach Sarvam before the boundary, or the
-            # final transcript is cut short.
-            await self._flush_audio_buffer()
-            await self._send_json({"event": "speech_end"})
-
-    async def run_stt(self, audio: bytes) -> AsyncGenerator[Frame | None, None]:
-        """Buffer and send raw audio bytes to Sarvam as base64-encoded JSON messages."""
-        if not audio:
-            yield None
-            return
-        if not self._is_websocket_open():
-            yield None
-            return
-
-        self._audio_buffer.extend(audio)
-        chunk_size = self._chunk_size_bytes()
-        while len(self._audio_buffer) >= chunk_size:
-            chunk = bytes(self._audio_buffer[:chunk_size])
-            del self._audio_buffer[:chunk_size]
-            try:
-                await self._send_audio_chunk(chunk)
-            except Exception as e:
-                await self.push_error(
-                    error_msg=f"Sarvam realtime STT send failed: {e}", exception=e
-                )
-                break
-
-        yield None
-
-    def _build_ws_url(self) -> str:
-        """Build the Sarvam realtime websocket URL."""
-        params = self._query_params()
-        return f"{self._base_url}?{urlencode(params)}"
-
-    async def _connect(self):
-        """Connect to Sarvam realtime and start receive task."""
-        await super()._connect()
-        await self._connect_websocket()
-        if self._websocket and not self._receive_task:
-            self._receive_task = self.create_task(self._receive_task_handler(self._report_error))
-
-    async def _disconnect(self):
-        """Disconnect from Sarvam realtime."""
-        await super()._disconnect()
-        if self._websocket and self._websocket.state is State.OPEN:
-            await self._flush_audio_buffer()
-            try:
-                await self._send_json({"event": "end"})
-            except Exception as e:
-                logger.debug(f"{self} error sending Sarvam end event: {e}")
-
-        if self._receive_task:
-            await self.cancel_task(self._receive_task)
-            self._receive_task = None
-
-        await self._disconnect_websocket()
-
-    async def _connect_websocket(self):
-        """Open the Sarvam realtime websocket."""
-        try:
-            if self._websocket and self._websocket.state is State.OPEN:
-                return
-
-            url = self._build_ws_url()
-            headers = {"API-SUBSCRIPTION-KEY": self._api_key}
-            # A fresh socket starts in whatever mode the query string asked for.
-            self._effective_endpointing = assert_given(self._settings.endpointing)
-            self._pending_endpointing = None
-            logger.debug(f"Connecting to Sarvam realtime STT WebSocket: {url}")
-            self._websocket = await websocket_connect(
-                url,
-                additional_headers=headers,
-                user_agent_header=sdk_headers()["User-Agent"],
-            )
-            await self._call_event_handler("on_connected")
-        except Exception as e:
-            self._websocket = None
-            await self.push_error(
-                error_msg=f"Unable to connect to Sarvam realtime STT: {e}", exception=e
-            )
-            await self._call_event_handler("on_connection_error", str(e))
-
-    async def _disconnect_websocket(self):
-        """Close the active websocket."""
-        try:
-            if self._websocket:
-                await self._websocket.close()
-        except Exception as e:
-            await self.push_error(
-                error_msg=f"Error closing Sarvam realtime STT websocket: {e}", exception=e
-            )
-        finally:
-            self._websocket = None
-            await self._call_event_handler("on_disconnected")
-
-    async def _receive_task_handler(self, report_error: Callable[[ErrorFrame], Awaitable[None]]):
-        """Close out the active utterance once the receive loop is done.
-
-        Reconnection is disabled, so the loop exiting means no further server
-        event can arrive. An utterance still open at that point would leave
-        downstream turn aggregation waiting on a boundary that is never coming.
-        Cancellation is left alone: that only happens during an intentional
-        disconnect, where teardown is already under way.
-        """
-        await super()._receive_task_handler(report_error)
-        await self._complete_active_utterance()
-
-    async def _receive_messages(self):
-        """Receive Sarvam realtime server events."""
-        if not self._websocket:
-            raise Exception("Websocket not connected")
-        async for message in self._websocket:
-            if not isinstance(message, str):
-                logger.trace(f"{self} ignored non-text Sarvam server message")
-                continue
-            try:
-                data = json.loads(message)
-            except json.JSONDecodeError:
-                logger.warning(f"{self} received non-JSON Sarvam message: {message}")
-                continue
-            await self._handle_message(data)
-
-    async def _handle_message(self, message: dict[str, Any]):
-        """Handle a parsed Sarvam realtime server event."""
-        event = message.get("event")
-        if event == "session.begin":
-            self._request_id = message.get("request_id")
-            logger.info(f"{self} Sarvam realtime session.begin request_id={self._request_id}")
-        # Only the endpointer in charge gets to set turn boundaries. Under
-        # manual endpointing the pipeline owns them, so server VAD telemetry
-        # would compete with the boundaries it is already producing.
-        elif event == "vad.speech_start" and self._effective_endpointing == "vad":
-            await self._handle_speech_start(message)
-        elif event == "vad.speech_end" and self._effective_endpointing == "vad":
-            await self._handle_speech_end(message)
-        elif event == "transcript.partial":
-            await self._handle_partial_transcript(message)
-        elif event == "transcript.final":
-            await self._handle_final_transcript(message)
-        elif event == "session.end":
-            await self._handle_session_end(message)
-        elif event == "config.updated":
-            await self._handle_config_updated(message)
-        elif event == "error":
-            await self._handle_error(message)
-        elif event == "pong":
-            # Answers our keepalive ping; the reply itself is the liveness proof.
-            logger.trace(f"{self} Sarvam realtime pong")
-        else:
-            logger.trace(f"{self} unhandled Sarvam realtime event: {message}")
-
-    async def update_config(self, **fields: Any):
-        """Send a live Sarvam ``config.update`` message without reconnecting.
-
-        Args:
-            fields: Runtime-updatable Sarvam config values. Connection-only
-                values are rejected, since Sarvam only reads them from the
-                connection query string.
-
-        Raises:
-            ValueError: If a connection-only or invalid value is supplied.
-        """
-        if not fields:
-            return
-        self._validate_config_update(fields)
-        await self._send_config_update(fields)
-
-    async def _send_config_update(self, fields: dict[str, Any]):
-        """Send an already-validated ``config.update`` and record what was sent."""
-        if not await self._send_json({"event": "config.update", **fields}):
-            return
-        # The store has to follow what the server was told, or a later delta
-        # diffs against a stale value and skips an update the server needs.
-        self._settings.apply_update(self.Settings(**fields))
-        if "endpointing" in fields:
-            # Boundary-gated: stay in the current mode until Sarvam acks.
-            self._pending_endpointing = fields["endpointing"]
-
-    async def _handle_config_updated(self, message: dict[str, Any]):
-        """Promote a pending endpointing switch once Sarvam acknowledges it."""
-        logger.trace(f"{self} Sarvam realtime acknowledgement: {message}")
-        if self._pending_endpointing is None:
-            return
-        applied = _applied_keys(message)
-        if applied is not None and "endpointing" not in applied:
-            # Someone else's acknowledgement; keep waiting for ours.
-            return
-        self._effective_endpointing = self._pending_endpointing
-        self._pending_endpointing = None
-        if self._effective_endpointing == "vad":
-            # Sarvam endpoints from here on, so the aggregator has to defer to
-            # the turn frames this service emits rather than keep running the
-            # detection it started the session with.
-            await self.broadcast_service_metadata()
-
-    async def _update_settings(self, delta: Settings) -> dict[str, Any]:
-        """Apply runtime settings and send supported fields via ``config.update``."""
-        delta = self._with_derived_language_code(delta)
-        proposed = self._settings.copy()
-        proposed.apply_update(delta)
-        self._validate_settings(proposed)
-
-        payload = {
-            name: getattr(proposed, name)
-            for name in _RUNTIME_CONFIG_FIELDS
-            if is_given(getattr(proposed, name))
-            and getattr(proposed, name) != getattr(self._settings, name)
-        }
-        # The wire guards compare a requested value against the live one, so
-        # they have to run while the store still holds it.
-        if payload:
-            self._validate_config_update(payload)
-
-        changed = await super()._update_settings(delta)
-        if not changed:
-            return changed
-
-        # `language` reaches Sarvam as `language_code`, so it is never unhandled.
-        unsupported = set(changed) - _RUNTIME_CONFIG_FIELDS - {"language"}
-        if unsupported:
-            self._warn_unhandled_updated_settings({key: changed[key] for key in unsupported})
-
-        if payload:
-            await self._send_config_update(payload)
-        return changed
-
-    @staticmethod
-    def _with_derived_language_code(delta: Settings) -> Settings:
-        """Fill in ``language_code`` from a ``language`` delta.
-
-        Mirrors the constructor: an explicit ``language_code`` wins, since it
-        also expresses ``auto``, which has no :class:`Language` equivalent.
-        """
-        if is_given(delta.language_code):
-            return delta
-        language = _as_language(delta.language)
-        if language is None:
-            return delta
-        derived = delta.copy()
-        derived.language_code = language_to_sarvam_realtime_language(language)
-        return derived
-
-    async def _handle_speech_start(self, message: dict[str, Any]):
-        if self._provider_speech_active:
-            return
-        self._provider_speech_active = True
-        self._speech_end_wall_time = None
-        self._speech_end_audio_position_s = None
-        await self.broadcast_frame(UserStartedSpeakingFrame)
-        if self._should_interrupt:
-            await self.broadcast_interruption()
-
-    async def _handle_speech_end(self, message: dict[str, Any]):
-        await self._complete_active_utterance()
-
-    async def _complete_active_utterance(self):
-        """End the in-flight utterance.
-
-        Runs on ``vad.speech_end`` and on a session ending mid-utterance. Without
-        the latter, downstream turn aggregation would wait forever for a boundary
-        the server is never going to send.
-        """
-        if not self._provider_speech_active:
-            return
-        self._provider_speech_active = False
-        self._speech_end_wall_time = time.time()
-        self._speech_end_audio_position_s = self._duration_for_bytes(self._audio_position_bytes)
-        await self.broadcast_frame(UserStoppedSpeakingFrame)
-        await self.start_ttfb_metrics(start_time=self._speech_end_wall_time)
-
-    async def _handle_partial_transcript(self, message: dict[str, Any]):
-        text = (message.get("text") or "").strip()
-        if not text:
-            return
-        result = self._result_payload(message)
-        language = self._language_for_frame(message.get("language"))
-        await self.push_frame(
-            InterimTranscriptionFrame(
-                text,
-                self._user_id,
-                time_now_iso8601(),
-                language,
-                result=result,
-            )
-        )
-
-    async def _handle_final_transcript(self, message: dict[str, Any]):
-        # Report usage before the transcription frame so tracing can attach it
-        # to the STT span the frame closes. A blank final still consumed audio.
-        await self.emit_stt_usage_metrics()
-        text = (message.get("text") or "").strip()
-        if text:
-            language = self._language_for_frame(message.get("language"))
-            result = self._result_payload(message)
-            result["speech_end_audio_position_s"] = self._speech_end_audio_position_s
-            await self.push_frame(
-                TranscriptionFrame(
-                    text,
-                    self._user_id,
-                    time_now_iso8601(),
-                    language,
-                    result=result,
-                    finalized=True,
-                )
-            )
-            await self._trace_transcription(text, True, language)
-
-    async def _handle_session_end(self, message: dict[str, Any]):
-        if message.get("request_id"):
-            self._request_id = message.get("request_id")
-        await self._complete_active_utterance()
-
-    async def _handle_error(self, message: dict[str, Any]):
-        await self.push_error(
-            error_msg=f"Sarvam realtime STT error: {json.dumps(message, ensure_ascii=False)}",
-        )
-
-    async def _send_audio_chunk(self, chunk: bytes):
-        if not self._websocket:
-            raise RuntimeError("WebSocket not connected")
-        sent = await self._send_json(
-            {"event": "audio_input", "audio": base64.b64encode(chunk).decode("utf-8")}
-        )
-        if not sent:
-            return
-        self._audio_position_bytes += len(chunk)
-
-    async def _flush_audio_buffer(self):
-        if not self._audio_buffer:
-            return
-        chunk = bytes(self._audio_buffer)
-        self._audio_buffer.clear()
-        if not self._is_websocket_open():
-            return
-        try:
-            await self._send_audio_chunk(chunk)
-        except Exception as e:
-            # Late audio on a socket the server already dropped must not break
-            # the turn boundary or teardown.
-            await self.push_error(error_msg=f"Sarvam realtime STT flush failed: {e}", exception=e)
-
-    async def _send_keepalive(self, silence: bytes):
-        """Hold the connection open with Sarvam's ping event.
-
-        Args:
-            silence: Silent PCM audio bytes, unused. The socket only accepts
-                JSON events, and Sarvam answers a ping with a pong rather than
-                reading padding audio.
-        """
-        await self._send_json({"event": "ping"})
-
-    async def _send_json(self, payload: dict[str, Any]) -> bool:
-        """Send a JSON event, reporting whether it reached the socket."""
-        if not self._is_websocket_open():
-            return False
-        assert self._websocket is not None
-        await self._websocket.send(json.dumps(payload))
-        return True
-
-    def _query_params(self) -> dict[str, Any]:
-        params: dict[str, Any] = {
-            "language_code": self._settings.language_code,
-            "stream_type": self._settings.stream_type,
-            "endpointing": self._settings.endpointing,
-            "encoding": "linear16",
-            "sample_rate": self.sample_rate,
-            "model": self._settings.model,
-            "mode": self._settings.mode,
-            "return_timestamps": str(self._settings.return_timestamps).lower(),
-        }
-        optional: dict[str, Any] = {
-            "prompt": self._settings.prompt,
-            "lid_gate_seconds": self._settings.lid_gate_seconds,
-            "lid_confidence_threshold": self._settings.lid_confidence_threshold,
-        }
-        if self._settings.endpointing == "vad":
-            optional.update(
-                {name: getattr(self._settings, name) for name in _VAD_TUNING_FIELDS},
-            )
-        params.update(
-            {key: value for key, value in optional.items() if is_given(value) and value is not None}
-        )
-        return params
-
-    def _bytes_per_second(self) -> int:
-        return self.sample_rate * 2
-
-    def _chunk_size_bytes(self) -> int:
-        return int(self._bytes_per_second() * (_CLIENT_CHUNK_MS / 1000))
-
-    def _duration_for_bytes(self, byte_count: int) -> float:
-        bytes_per_second = self._bytes_per_second()
-        if bytes_per_second <= 0:
-            return 0.0
-        return byte_count / bytes_per_second
-
-    def _result_payload(self, message: dict[str, Any]) -> dict[str, Any]:
-        payload = dict(message)
-        if self._request_id is not None:
-            payload.setdefault("request_id", self._request_id)
-        confidence = payload.get("confidence")
-        if isinstance(confidence, bool) or not isinstance(confidence, (int, float)):
-            # `language_confidence` is a LID score and never a recognition score.
-            payload["confidence"] = 1.0
-        return payload
-
-    def _language_for_frame(self, raw_language: str | None = None) -> Language | None:
-        configured_language_code = assert_given(self._settings.language_code)
-        language_code = self._normalize_language_code(raw_language or configured_language_code)
-        if language_code == "auto":
-            return None
-        try:
-            return Language(language_code)
-        except ValueError:
-            return None
-
-    def _normalize_language_code(self, language_code: str | None) -> str:
-        if not language_code:
-            return assert_given(self._settings.language_code)
-        if "-" not in language_code and language_code != "auto":
-            configured = assert_given(self._settings.language_code)
-            if configured != "auto" and configured.startswith(f"{language_code}-"):
-                return configured
-            return _SHORT_LANGUAGE_DEFAULTS.get(language_code, language_code)
-        return language_code
-
-    def _is_websocket_open(self) -> bool:
-        return self._websocket is not None and self._websocket.state is State.OPEN
-
-    def _resolved_sample_rate_error(self) -> str | None:
-        """Describe why the rate actually in use is unusable, if it is."""
-        if self.sample_rate in SUPPORTED_SAMPLE_RATES:
-            return None
-        allowed = ", ".join(str(rate) for rate in sorted(SUPPORTED_SAMPLE_RATES))
-        return f"Unsupported sample_rate '{self.sample_rate}'. Allowed values: {allowed}."
-
-    def _validate_config_update(self, fields: dict[str, Any]):
-        connection_only = sorted(_CONNECTION_ONLY_FIELDS.intersection(fields))
-        if connection_only:
-            names = ", ".join(connection_only)
-            raise ValueError(
-                f"{names} cannot be changed mid-session; they are connection "
-                "parameters. Open a new stream instead."
-            )
-        unknown = sorted(set(fields) - _RUNTIME_CONFIG_FIELDS)
-        if unknown:
-            names = ", ".join(unknown)
-            allowed = ", ".join(sorted(_RUNTIME_CONFIG_FIELDS))
-            raise ValueError(f"Unsupported config.update field(s) {names}. Allowed: {allowed}.")
-        if "stream_type" in fields:
-            current = assert_given(self._settings.stream_type)
-            requested = fields["stream_type"]
-            if current == "simulated" or requested == "simulated":
-                raise ValueError("Changing to or from stream_type='simulated' is not supported.")
-        proposed = self._settings.copy()
-        proposed.apply_update(self.Settings(**fields))
-        self._validate_settings(proposed)
-
-    @staticmethod
-    def _validate_settings(settings: Settings):
-        model = assert_given(settings.model)
-        if model != _REALTIME_MODEL:
-            raise ValueError(f"Unsupported model '{model}'. Only '{_REALTIME_MODEL}' is supported.")
-
-        language_code = assert_given(settings.language_code)
-        if language_code not in SUPPORTED_LANGUAGES:
-            allowed = ", ".join(sorted(SUPPORTED_LANGUAGES))
-            raise ValueError(
-                f"Unsupported language_code '{language_code}'. Allowed values: {allowed}."
-            )
-
-        stream_type = assert_given(settings.stream_type)
-        if stream_type not in SUPPORTED_STREAM_TYPES:
-            allowed = ", ".join(sorted(SUPPORTED_STREAM_TYPES))
-            raise ValueError(f"Unsupported stream_type '{stream_type}'. Allowed values: {allowed}.")
-
-        endpointing = assert_given(settings.endpointing)
-        if endpointing not in SUPPORTED_ENDPOINTING:
-            allowed = ", ".join(sorted(SUPPORTED_ENDPOINTING))
-            raise ValueError(f"Unsupported endpointing '{endpointing}'. Allowed values: {allowed}.")
-
-        sample_rate = assert_given(settings.sample_rate)
-        if sample_rate is not None and sample_rate not in SUPPORTED_SAMPLE_RATES:
-            allowed = ", ".join(str(rate) for rate in sorted(SUPPORTED_SAMPLE_RATES))
-            raise ValueError(f"Unsupported sample_rate '{sample_rate}'. Allowed values: {allowed}.")
-
-        mode = assert_given(settings.mode)
-        if mode not in SUPPORTED_MODES:
-            allowed = ", ".join(sorted(SUPPORTED_MODES))
-            raise ValueError(f"Unsupported mode '{mode}'. Allowed values: {allowed}.")
-
-        _validate_optional_range("threshold", settings.threshold, minimum=0.0, maximum=1.0)
-        _validate_optional_minimum("prefix_padding_ms", settings.prefix_padding_ms, minimum=0)
-        _validate_optional_minimum("silence_duration_ms", settings.silence_duration_ms, minimum=0)
-        _validate_optional_minimum(
-            "min_speech_duration_ms", settings.min_speech_duration_ms, minimum=0
-        )
-        _validate_optional_minimum("lid_gate_seconds", settings.lid_gate_seconds, minimum=0.0)
-        _validate_optional_range(
-            "lid_confidence_threshold",
-            settings.lid_confidence_threshold,
-            minimum=0.0,
-            maximum=1.0,
-        )
-
-    @traced_stt
-    async def _trace_transcription(
-        self, transcript: str, is_final: bool, language: Language | None = None
-    ):
-        """Record transcription event for tracing."""
-        pass
-
-
-def _as_language(value: Any) -> Language | None:
-    """Coerce a settings ``language`` value to a ``Language``, or ``None``."""
-    if isinstance(value, Language):
-        return value
-    if not isinstance(value, str):
-        return None
-    try:
-        return Language(value)
-    except ValueError:
-        return None
-
-
-def _applied_keys(message: dict[str, Any]) -> set[str] | None:
-    """Extract the field names from a ``config.updated`` payload.
-
-    Sarvam reports entries as ``"key=value"``. Returns ``None`` when the payload
-    carries no usable list, which callers treat as "assume it was mine".
-    """
-    applied = message.get("applied")
-    if not isinstance(applied, list):
-        return None
-    keys = {str(item).split("=", maxsplit=1)[0].strip() for item in applied}
-    return keys or None
-
-
-def _validate_optional_range(
-    field_name: str,
-    value: float | int | None | NotGiven,
-    *,
-    minimum: float,
-    maximum: float,
-):
-    if not is_given(value) or value is None:
-        return
-    if value < minimum or value > maximum:
-        raise ValueError(f"{field_name} must be between {minimum} and {maximum}.")
-
-
-def _validate_optional_minimum(
-    field_name: str,
-    value: float | int | None | NotGiven,
-    *,
-    minimum: float,
-):
-    if not is_given(value) or value is None:
-        return
-    if value < minimum:
-        raise ValueError(f"{field_name} must be greater than or equal to {minimum}.")
